// src/app/models/address.ts
export class UserAddress {

    addressId: number;
    city: string;
    // country: string;
    // geolocation: null
    landmark: string;
    name: string;
    pincode: string;
    flatNo: string;
    apartment: string;
    street: string;
    area: string
    primaryAddress: boolean;
    number: string;
    // state: null
    type: string;
}